Select device by usage
There are two HID devices exposed, and these are returned in a
non-deterministic order. The code doesn't make any attempt to grab the
correct device, so sometimes if the 'keyboard' device is grabbed, the
code locks up.

Enumerate all the devices and get the correct one based on the
usage_page.